+++ modules/server/src/main/java/io/jpom/model/data/NodeVersionModel.java
@@ -12,6 +12,8 @@ public class NodeVersionModel extends BaseModel {
      */
     private String group;
 
+
+
     public String getVersion() {
         return version;
     }